Automatic icon positioning
This commit introduces automatic icon positioning on a grid into kdesk.

You can turn it on by putting the following into the icon config:

Relative-to: grid
X: <grid-location-hint>
Y: <grid-location-hint>

kdesk will try to satisfy the field based on your location hints, but
in cases it cannot, it will position the icon automatically. In case
you don't care about the position of the icon at all, just set X and Y
to 'auto'.

Note that the attributes Width and Height are IGNORED when the icon
is put into the grid! Both of them are forced to 128, so the grid
is uniform.

Related to: https://github.com/KanoComputing/peldins/issues/716

TODO: We might want to rename the attribute Relative-to to something
different in the future.